Refactored and expanded unit tests
