+++ tests/unittests/test_dto_utils.py
@@ -1,3 +1,4 @@
+# pylint: disable="missing-class-docstring", "missing-function-docstring"
 import pathlib
 import sys
 import unittest
